fix auto executor selection
fix #223

This PR fixes that if the user desabled in CMake all exectors the
automatic executor selection for function where the executor is optional
was not working.
The root of the issue was that the code used the list of executors which
was designed for the tests and examples and can be managed by CMake
flags.

- introduce `alpaka::allExecutors`
- move `executor.hpp` to the root folder to signal the user that the
  content can be used in normal applications
- move methods to select avaiable backends and executors to `onHost`
- move `enabledExecutors` to `onhost::example` to clearly show the side
  effects the usage has